+++ cocos/2d/ZipUtils.cpp
@@ -591,7 +591,7 @@ bool ZipFile::fileExists(const std::string &fileName) const
     return ret;
 }
 
-unsigned char *ZipFile::getFileData(const std::string &fileName, long *size)
+unsigned char *ZipFile::getFileData(const std::string &fileName, size_t *size)
 {
     unsigned char * buffer = NULL;
     if (size)
